temp commit, test vim-plugs

+++ setup.sh
@@ -1,4 +1,9 @@
 git submodule init && git submodule update
+
+if [ -z "${PLUGIN_MANAGER}" ]; then
+    PLUGIN_MANAGER="vim-plug"
+fi
+
 if [ ! -e .swp ]; then
     mkdir .swp
 fi
@@ -9,10 +14,19 @@ fi
 if [ ! -e bundle ]; then
     mkdir bundle
 fi
-cd bundle
-if [ ! -e vundle ]; then
+
+pushd bundle
+if [ ! -e vundle -a "${PLUGIN_MANAGER}" == "vundle" ]; then
     ln -s ../ext-repo/vundle vundle
 fi
+popd
+
+if [ "${PLUGIN_MANAGER}" == "vim-plug" ]; then
+    pushd autoload
+    ln -s ../ext-repo/vim-plug vim-plug/plug.vim .
+    popd
+fi
+
 cd ..
 cp available_plugins.vim installed_plugins.vim
